+++ app/db/crud/package_crud.py
@@ -1,6 +1,6 @@
 from fastapi import HTTPException, status
 from sqlalchemy.orm import Session
-from models import PackageModel
+from models import PackageModel, PackageReservation
 from schemas import PackageSchema
 from db.crud.agency_crud import get_agency
 from db.crud.extended_excursion_crud import get_extended_excursion
@@ -46,13 +46,50 @@ def delete_package(db: Session, package_delete: PackageSchema):
     
     package = get_package(db, package_delete.id, package_delete.agency_id, package_delete.extended_excursion_id)
     if package is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Package association not found")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Package not found")
 
+    package_in_reservation = db.query(PackageReservation).filter(PackageReservation.package_id == package_delete.id).first()
+    if package_in_reservation is not None:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Can't delete this Package because is is ivolved in a Package Reservation")
 
     db.delete(package)
     db.commit()
 
     return "Success"
+
+def update_package(db: Session, package_update: PackageSchema):
+
+    agency = get_agency(db, package_update.agency_id)
+    if agency is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Agency not found")
+    
+    extended_excursion = get_extended_excursion(db, package_update.extended_excursion_id)
+    if extended_excursion is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Extended excursion not found")
+    
+    package = get_package(db, package_update.id, package_update.agency_id, package_update.extended_excursion_id)
+    if package is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Package not found")
+    
+    package_in_reservation = db.query(PackageReservation).filter(PackageReservation.package_id == package_update.id).first()
+    if package_in_reservation is not None:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Can't update this Package because is is ivolved in a Package Reservation")
+    
+    if package_update.price is not None:
+        package.price = package_update.price
+    if package_update.description is not None:
+        package.description = package_update.description
+    if package_update.duration is not None:
+        package.duration = package_update.duration
+    if package_update.photo_url is not None:
+        package.photo_url = package_update.photo_url
+    
+
+
+    db.commit()
+
+    return "Success"
+
 
 def toModel(schema:PackageSchema) -> PackageModel:
     return PackageModel(
